fix: apply gitignore and delete IDE files
It is a bad idea to keep user and IDE specific files like .idea in the project. Every user might use different IDE and files of one user could interfere with the system of the other one. Gitignore - is a standard way to keep project clean of such technical files and folders.